Move all apps to app dir
