+++ src/apps/common/wagtail_hooks.py
@@ -1,5 +1,5 @@
 from wagtail.snippets.models import register_snippet
 
-from src.common import models
+from src.apps.common import models
 
 register_snippet(models.Link)
